conferences/requirements.txt
Django==1.9
factory-boy==2.6.0
fake-factory==0.5.3
MySQL-python==1.2.5
python-dateutil==2.4.2
six==1.10.0
wheel==0.24.0
